:bug: Fix type issue

+++ src/rt/react.c
@@ -33,7 +33,7 @@ void Unwrap(char *str, const char *str0)
 {
     int             i, j = 0;
 
-    for (i = 0; i < strlen(str0); i++)
+    for (i = 0; i < (int)strlen(str0); i++)
     {
         if (str0[i] != '\'')
         {
